LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerlock/pointerevent_getCoalescedEvents_when_pointerlocked-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
PointerMove getCoalescedEvent in locked state test

Follow the test instructions with mouse. If you don't have the device skip it.

Test Description: This test checks if pointerevent.getCoalescedEvent work correctly when pointer is locked.
Press left button down on the green rectangle to lock pointer.
Move the mouse
Test passes if the proper behavior of the events is observed.


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN mouse pointermove getCoalescedEvents when lock test
